refactor: split lib/analytics.ts into utils and services
Separate device detection (utils/device.ts), geolocation IO
(services/geolocation.ts), and tracking (services/analytics.ts). Analytics.tsx
now imports from the services layer; lib/analytics.ts is removed, as it had a
single consumer and needs no re-export shim. Adds an explicit GeoData
interface, replacing the previously untyped emptyGeo literal and any-typed
getGeolocation return.

Closes #1470
